requirements.txt
pandas==2.2.2
numpy==2.0.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-instrumentation-logging==1.31.1
